Make receive queues negative and zero based.
Descriptor queue lengths (qlen) is how much descriptors are owned by the
card.  So empty rx queue is always equal in size to the maximum queue
length (minus 1). In other words, it is how much descriptors are free
for the card to fill new packets. But human would expect 'queue length'
to be how much packets are outstanding waiting for processing. Thus,
I subtract hw qlen from queue size (and add 1 to make it 0-based) to
make it negative waiting queue length.  Negative, to be fair that
meaning is not really wait queue but negative of free space. Which in
most cases should be equal to wait queue length (but may not).

This will also make qlen row disappearing when everything is 0 (which is
in normal situation).